--- Dayana/Shared/Persistence/Models/Blog/Filters/PostCategoryFilters/PostCategorySortBy.cs
@@ -1,7 +1,0 @@
-﻿namespace Dayana.Shared.Persistence.Models.Blog.Filters.BlogPosts.PostCategoryFilters;
-
-public enum PostCategorySortBy
-{
-    CreationDate = 12,
-    CreationDateDescending = 13,
-}
