--- Dayana/Shared/Persistence/Models/Blog/Filters/PostFilters/PostSortBy.cs
@@ -1,7 +1,0 @@
-﻿namespace Dayana.Shared.Persistence.Models.Blog.Filters.BlogPosts.PostFilters;
-
-public enum PostSortBy
-{
-    CreationDate = 12,
-    CreationDateDescending = 13,
-}
